fix: change Resource.Size from *int64 to *float64 per spec
The MCP spec defines size as 'number', which in JSON can be either
integer or floating-point. Using *int64 caused a JSON decode error
when a server returned a float value like 12345.0.

Switch to *float64 to accept both formats without error. The field
is currently unused in application code, so there is no caller impact.

+++ internal/mcp/types.go
@@ -301,7 +301,7 @@ type Resource struct {
 	MIMEType    string       `json:"mimeType,omitempty"`
 	Annotations *Annotations `json:"annotations,omitempty"`
 	Icons       []Icon       `json:"icons,omitempty"`
-	Size        *int64       `json:"size,omitempty"`
+	Size        *float64     `json:"size,omitempty"`
 	Meta        Meta         `json:"_meta,omitempty"`
 }
